+++ ios/AppMetricaPushPlugin/AMPFPigeon.h
@@ -34,6 +34,7 @@ NSObject<FlutterMessageCodec> *AMPFAppMetricaPushPigeonGetCodec(void);
 
 @protocol AMPFAppMetricaPushPigeon
 - (void)activateWithError:(FlutterError *_Nullable *_Nonnull)error;
+- (void)activateWithProvidersProviders:(NSArray<NSString *> *)providers error:(FlutterError *_Nullable *_Nonnull)error;
 - (void)saveAppMetricaConfigConfig:(NSString *)config error:(FlutterError *_Nullable *_Nonnull)error;
 - (void)getTokensWithCompletion:(void (^)(NSDictionary<NSString *, NSString *> *_Nullable, FlutterError *_Nullable))completion;
 - (void)getLaunchPushInfoWithCompletion:(void (^)(AMPFAppMetricaPushInfoPigeon *_Nullable, FlutterError *_Nullable))completion;