+++ lib/appmetrica_push_plugin.dart
@@ -1,3 +1,7 @@
 library appmetrica_push;
 
 export 'src/appmetrica_push.dart' show AppMetricaPush;
+
+export 'src/push_provider.dart' show PushProvider;
+
+export 'src/firebase_push_provider.dart' show FirebasePushProvider;