+++ lib/src/firebase_push_provider.dart
@@ -1,0 +1,6 @@
+import 'push_provider.dart';
+
+class FirebasePushProvider extends PushProvider {
+  @override
+  String get nativeFactoryClass => "io.appmetrica.analytics.push.flutter.FirebasePushProviderFactory";
+}